+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/deepseek
 
+## 2.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [bb36798]
+  - @ai-sdk/provider@3.0.0-beta.15
+  - @ai-sdk/openai-compatible@2.0.0-beta.32
+  - @ai-sdk/provider-utils@4.0.0-beta.31
+
 ## 2.0.0-beta.31
 
 ### Patch Changes