+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/rsc
 
+## 2.0.0-beta.93
+
+### Patch Changes
+
+- Updated dependencies [bb36798]
+  - @ai-sdk/provider@3.0.0-beta.15
+  - ai@6.0.0-beta.93
+  - @ai-sdk/provider-utils@4.0.0-beta.31
+
 ## 2.0.0-beta.92
 
 ### Patch Changes