+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.93
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [97b1d77]
   - ai@6.0.0-beta.92